refactor: separating token actions in different files
